Updated authorization in methods

+++ src/users/users.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   ParseIntPipe,
@@ -10,45 +11,71 @@ import {
   Query,
   ValidationPipe,
 } from '@nestjs/common';
-import { Public } from 'src/auth/decorator';
+import { GetUser, Roles } from 'src/auth/decorator';
 import { UsersService } from './users.service';
 import { CreateUserDto, GetUsersDto, UpdateUserDto } from './dto';
+import { Role } from '../auth/enums/role.enum';
 
 @Controller('users')
 export class UsersController {
   constructor(private usersService: UsersService) {}
 
-  @Public()
+  @Roles(Role.ADMIN)
   @Get()
   getUsers(@Query() query: GetUsersDto) {
     return this.usersService.getUsers(query);
   }
 
-  @Public()
+  @Roles(Role.ADMIN)
   @Post()
   createUser(@Body() dto: CreateUserDto) {
     return this.usersService.createUser(dto);
   }
 
-  @Public()
+  @Roles(Role.USER, Role.ADMIN)
   @Get(':id')
-  getUser(@Param('id', ParseIntPipe) id: number) {
+  getUser(
+    @Param('id', ParseIntPipe) id: number,
+    @GetUser('id') userId: number,
+    @GetUser('role') userRole: Role,
+  ) {
+    // Se verifica que el usuario está autorizado
+    if (userRole === Role.USER && userId !== id) {
+      throw new ForbiddenException('You are not allowed to access this user.');
+    }
+
     return this.usersService.getUser(id);
   }
 
-  @Public()
+  @Roles(Role.USER, Role.ADMIN)
   @Put(':id')
   updateUser(
     @Param('id', ParseIntPipe) id: number,
+    @GetUser('id') userId: number,
+    @GetUser('role') userRole: Role,
     @Body(new ValidationPipe({ skipMissingProperties: true }))
     dto: UpdateUserDto,
   ) {
+    // Se verifica que el usuario está autorizado
+    if (userRole === Role.USER && userId !== id) {
+      throw new ForbiddenException('You are not allowed to access this user.');
+    }
+
     return this.usersService.updateUser(id, dto);
   }
 
-  @Public()
+  @Roles(Role.USER, Role.ADMIN)
   @Delete(':id')
-  deleteUser(@Param('id', ParseIntPipe) id: number) {
+  deleteUser(
+    @Param('id', ParseIntPipe) id: number,
+    @GetUser('id') userId: number,
+    @GetUser('role') userRole: Role,
+  ) {
+    // Se verifica que el usuario está autorizado
+    if (userRole === Role.USER && userId !== id) {
+      throw new ForbiddenException('You are not allowed to access this user.');
+    }
+
     return this.usersService.deleteUser(id);
   }
 }
